Big reorganise of functions.
