+++ frontend/mainWindow.cpp
@@ -3,37 +3,13 @@
 
 #include "mainWindow.h"
 #include "ui_mainWindow.h"
-#include <QChart>
-#include <QChartView>
-#include <QCheckBox>
-#include <QDialog>
-#include <QDialogButtonBox>
-#include <QDomDocument>
-#include <QFormLayout>
-#include <QInputDialog>
-#include <QLabel>
-#include <QMessageBox>
-#include <QNetworkReply>
 #include <QSettings>
 #include <QTimer>
-#include <QWidgetAction>
-#include <QtGui>
 
 MainWindow::MainWindow(QWidget *parent) : QMainWindow(parent)
 {
     ui_.setupUi(this);
-    initialiseElements();
 
-    QTimer *timer = new QTimer(this);
-    connect(timer, &QTimer::timeout, [=]() { checkForUpdates(); });
-    timer->start(30000);
-}
-
-MainWindow::~MainWindow() {}
-
-// Configure initial application state
-void MainWindow::initialiseElements()
-{
     auto instruments = getInstruments();
     fillInstruments(instruments);
 
@@ -96,7 +72,13 @@ void MainWindow::initialiseElements()
         auto *worker2 = new HttpRequestWorker(this);
         worker2->execute(input2);
     }
+
+    QTimer *timer = new QTimer(this);
+    connect(timer, &QTimer::timeout, [=]() { checkForUpdates(); });
+    timer->start(30000);
 }
+
+MainWindow::~MainWindow() {}
 
 // Sets cycle to most recently viewed
 void MainWindow::recentCycle()
@@ -169,105 +151,6 @@ void MainWindow::closeEvent(QCloseEvent *event)
     event->accept();
 }
 
-void MainWindow::massSearch(QString name, QString value)
-{
-    const char *prompt;
-    QString textInput;
-    QString text;
-    bool caseSensitivity = false;
-    name.append(": ");
-    // configure for ranges
-    if (name == "Run Range: ")
-        prompt = "StartRun-EndRun:";
-    else if (name == "Date Range: ")
-        prompt = "yyyy/mm/dd-yyyy/mm/dd:";
-    else
-        prompt = name.toUtf8();
-
-    if (name.contains("Range"))
-    {
-        QDialog dialog(this);
-        QFormLayout form(&dialog);
-
-        form.addRow(new QLabel(name));
-        auto *start = new QLineEdit(&dialog);
-        form.addRow("Start:", start);
-        auto *end = new QLineEdit(&dialog);
-        form.addRow("End:", end);
-
-        QDialogButtonBox buttonBox(QDialogButtonBox::Ok | QDialogButtonBox::Cancel, Qt::Horizontal, &dialog);
-        form.addRow(&buttonBox);
-        QObject::connect(&buttonBox, SIGNAL(accepted()), &dialog, SLOT(accept()));
-        QObject::connect(&buttonBox, SIGNAL(rejected()), &dialog, SLOT(reject()));
-
-        if (dialog.exec() == QDialog::Accepted)
-            textInput = start->text() + "-" + end->text();
-        else
-            return;
-    }
-    else
-    {
-        QDialog dialog(this);
-        QFormLayout form(&dialog);
-
-        form.addRow(new QLabel(name));
-        auto *input = new QLineEdit(&dialog);
-        form.addRow(name, input);
-        auto *caseSensitive = new QCheckBox(&dialog);
-        form.addRow("Case sensitive", caseSensitive);
-
-        QDialogButtonBox buttonBox(QDialogButtonBox::Ok | QDialogButtonBox::Cancel, Qt::Horizontal, &dialog);
-        form.addRow(&buttonBox);
-        QObject::connect(&buttonBox, SIGNAL(accepted()), &dialog, SLOT(accept()));
-        QObject::connect(&buttonBox, SIGNAL(rejected()), &dialog, SLOT(reject()));
-
-        if (dialog.exec() == QDialog::Accepted)
-        {
-            textInput = input->text();
-            caseSensitivity = caseSensitive->isChecked();
-        }
-        else
-            return;
-    }
-    text = name.append(textInput);
-    textInput.replace("/", ";");
-    if (textInput.isEmpty())
-        return;
-    for (auto tuple : cachedMassSearch_)
-    {
-        if (std::get<1>(tuple) == text)
-        {
-            for (QAction *action : cyclesMenu_->actions())
-            {
-                if (action->text() == "[" + std::get<1>(tuple) + "]")
-                    action->trigger();
-            }
-            setLoadScreen(true);
-            return;
-        }
-    }
-
-    // mass search for data
-    QString searchOptions;
-    QString sensitivityText = "caseSensitivity=";
-    sensitivityText.append(caseSensitivity ? "true" : "false");
-    searchOptions.append(sensitivityText);
-    QString url_str = "http://127.0.0.1:5000/getAllJournals/" + instName_ + "/" + value + "/" + textInput + "/" + searchOptions;
-    HttpRequestInput input(url_str);
-    auto *worker = new HttpRequestWorker(this);
-    connect(worker, SIGNAL(on_execution_finished(HttpRequestWorker *)), this, SLOT(handle_result_cycles(HttpRequestWorker *)));
-    worker->execute(input);
-
-    // configure caching
-    cachedMassSearch_.append(std::make_tuple(worker, text));
-
-    auto *action = new QAction("[" + text + "]", this);
-    connect(action, &QAction::triggered, [=]() { changeCycle("[" + text + "]"); });
-    cyclesMenu_->addAction(action);
-    ui_.cycleButton->setText("[" + text + "]");
-    setLoadScreen(true);
-}
-
 void MainWindow::keyPressEvent(QKeyEvent *event)
 {
     if (event->key() == Qt::Key_G && event->modifiers() == Qt::ControlModifier)
@@ -284,236 +167,6 @@ void MainWindow::keyPressEvent(QKeyEvent *event)
         updateSearch(searchString_);
         return;
     }
-}
-
-// Get instrument data from config file
-QList<std::tuple<QString, QString, QString>> MainWindow::getInstruments()
-{
-    QFile file(":/data/instrumentData.xml");
-    if (!file.exists())
-        return {};
-
-    file.open(QIODevice::ReadOnly);
-    QDomDocument dom;
-    dom.setContent(&file);
-    file.close();
-    auto rootelem = dom.documentElement();
-    auto nodelist = rootelem.elementsByTagName("inst");
-    auto headersList = rootelem.elementsByTagName("header");
-    headersMap_.clear();
-    QString header;
-    QString data;
-    for (auto i = 0; i < headersList.count(); i++)
-    {
-        header = headersList.item(i).toElement().attribute("name");
-        data = headersList.item(i).toElement().elementsByTagName("Data").item(0).toElement().text();
-        headersMap_[data] = header;
-    }
-
-    QList<std::tuple<QString, QString, QString>> instruments;
-    std::tuple<QString, QString, QString> instrument;
-    QDomNode node;
-    QDomElement elem;
-    for (auto i = 0; i < nodelist.count(); i++)
-    {
-        node = nodelist.item(i);
-        elem = node.toElement();
-        auto instrumentDisplayName = elem.elementsByTagName("displayName").item(0).toElement().text();
-        auto instrumentType = elem.elementsByTagName("type").item(0).toElement().text();
-        auto instrumentName = elem.attribute("name");
-        instruments.append(std::make_tuple(instrumentName, instrumentType, instrumentDisplayName));
-    }
-    return instruments;
-}
-
-QDomDocument MainWindow::getConfig()
-{
-    QDomDocument dom;
-    QSettings settings(QSettings::IniFormat, QSettings::UserScope, "ISIS", "jv2");
-    if (!settings.contains("tableConfig"))
-    {
-        QFile file(":/data/tableConfig.xml");
-        file.open(QIODevice::ReadOnly);
-        dom.setContent(&file);
-        file.close();
-    }
-    else
-        dom.setContent(settings.value("tableConfig", "fail").toString());
-    return dom;
-}
-
-// Get the desired fields and their titles
-std::vector<std::pair<QString, QString>> MainWindow::getFields(QString instrument, QString instType)
-{
-    std::vector<std::pair<QString, QString>> desiredInstFields;
-    QDomNodeList desiredInstrumentFields;
-    auto dom = getConfig();
-
-    std::pair<QString, QString> column;
-
-    auto rootelem = dom.documentElement();
-    auto instList = rootelem.elementsByTagName("inst");
-
-    for (auto i = 0; i < instList.count(); i++)
-    {
-        if (instList.item(i).toElement().attribute("name").toLower() == instrument)
-        {
-            desiredInstrumentFields = instList.item(i).toElement().elementsByTagName("Column");
-            break;
-        }
-    }
-    // If inst preferences blank
-    if (desiredInstrumentFields.isEmpty())
-    {
-        auto configDefault = rootelem.elementsByTagName(instType).item(0).toElement();
-        auto configDefaultFields = configDefault.elementsByTagName("Column");
-        // If config preferences blank
-        if (configDefaultFields.isEmpty())
-        {
-            QFile file(":/data/instrumentData.xml");
-            file.open(QIODevice::ReadOnly);
-            dom.setContent(&file);
-            file.close();
-            auto rootelem = dom.documentElement();
-            auto defaultColumns = rootelem.elementsByTagName(instType).item(0).toElement().elementsByTagName("Column");
-            // Get config preferences
-            for (int i = 0; i < defaultColumns.count(); i++)
-            {
-                // Get column index and title from xml
-                column.first = defaultColumns.item(i).toElement().elementsByTagName("Data").item(0).toElement().text();
-                column.second = defaultColumns.item(i).toElement().attribute("name");
-                desiredInstFields.push_back(column);
-            }
-            return desiredInstFields;
-        }
-        // Get config default
-        for (int i = 0; i < configDefaultFields.count(); i++)
-        {
-            column.first = configDefaultFields.item(i).toElement().elementsByTagName("Data").item(0).toElement().text();
-            column.second = configDefaultFields.item(i).toElement().attribute("name");
-            desiredInstFields.push_back(column);
-        }
-        return desiredInstFields;
-    }
-    // Get instrument preferences
-    for (int i = 0; i < desiredInstrumentFields.count(); i++)
-    {
-        column.first = desiredInstrumentFields.item(i).toElement().elementsByTagName("Data").item(0).toElement().text();
-        column.second = desiredInstrumentFields.item(i).toElement().attribute("name");
-        desiredInstFields.push_back(column);
-    }
-    return desiredInstFields;
-}
-
-void MainWindow::savePref()
-{
-    auto dom = getConfig();
-
-    auto rootelem = dom.documentElement();
-    auto nodelist = rootelem.elementsByTagName("inst");
-    // Get current table fields
-    QString currentFields;
-    int realIndex;
-    for (auto i = 0; i < ui_.runDataTable->horizontalHeader()->count(); ++i)
-    {
-        realIndex = ui_.runDataTable->horizontalHeader()->logicalIndex(i);
-        if (!ui_.runDataTable->isColumnHidden(realIndex))
-        {
-            currentFields += model_->headerData(realIndex, Qt::Horizontal, Qt::UserRole).toString();
-            currentFields += ",";
-            currentFields += model_->headerData(realIndex, Qt::Horizontal).toString();
-            currentFields += ",;";
-        }
-    }
-    currentFields.chop(1);
-
-    // Add preferences to xml file
-    QDomNode node;
-    QDomElement elem;
-    QDomElement columns;
-    for (auto i = 0; i < nodelist.count(); i++)
-    {
-        node = nodelist.item(i);
-        elem = node.toElement();
-        if (elem.attribute("name") == instName_)
-        {
-            auto oldColumns = elem.elementsByTagName("Columns");
-            if (!oldColumns.isEmpty())
-                elem.removeChild(elem.elementsByTagName("Columns").item(0));
-            columns = dom.createElement("Columns");
-            for (QString field : currentFields.split(";"))
-            {
-                auto preferredFieldsElem = dom.createElement("Column");
-                auto preferredFieldsDataElem = dom.createElement("Data");
-                preferredFieldsElem.setAttribute("name", field.split(",")[1]);
-                preferredFieldsDataElem.appendChild(dom.createTextNode(field.split(",")[0]));
-                preferredFieldsElem.appendChild(preferredFieldsDataElem);
-                columns.appendChild(preferredFieldsElem);
-            }
-            elem.appendChild(columns);
-        }
-    }
-    if (!dom.toByteArray().isEmpty())
-    {
-        QSettings settings(QSettings::IniFormat, QSettings::UserScope, "ISIS", "jv2");
-        settings.setValue("tableConfig", dom.toByteArray());
-    }
-}
-
-void MainWindow::clearPref()
-{
-    auto dom = getConfig();
-
-    auto rootelem = dom.documentElement();
-    auto nodelist = rootelem.elementsByTagName("inst");
-
-    // Clear preferences from xml file
-    QDomNode node;
-    QDomElement elem;
-    QDomElement columns;
-    for (auto i = 0; i < nodelist.count(); i++)
-    {
-        node = nodelist.item(i);
-        elem = node.toElement();
-        if (elem.attribute("name") == instName_)
-        {
-            auto oldColumns = elem.elementsByTagName("Columns");
-            if (!oldColumns.isEmpty())
-                elem.removeChild(elem.elementsByTagName("Columns").item(0));
-        }
-    }
-    if (!dom.toByteArray().isEmpty())
-    {
-        QSettings settings(QSettings::IniFormat, QSettings::UserScope, "ISIS", "jv2");
-        settings.setValue("tableConfig", dom.toByteArray());
-    }
-}
-
-void MainWindow::on_actionMountPoint_triggered()
-{
-    QString textInput = QInputDialog::getText(this, tr("Set Mount Point"), tr("Location:"), QLineEdit::Normal);
-    if (textInput.isEmpty())
-        return;
-
-    QSettings settings(QSettings::IniFormat, QSettings::UserScope, "ISIS", "jv2");
-    settings.setValue("mountPoint", textInput);
-
-    QString url_str = "http://127.0.0.1:5000/setRoot/";
-    url_str += textInput;
-    HttpRequestInput input(url_str);
-    auto *worker = new HttpRequestWorker(this);
-    worker->execute(input);
-}
-
-void MainWindow::on_actionClearMountPoint_triggered()
-{
-    QSettings settings(QSettings::IniFormat, QSettings::UserScope, "ISIS", "jv2");
-    settings.setValue("mountPoint", "");
-
-    QString url_str = "http://127.0.0.1:5000/setRoot/Default";
-    HttpRequestInput input(url_str);
-    auto *worker = new HttpRequestWorker(this);
-    worker->execute(input);
 }
 
 void MainWindow::setLoadScreen(bool state)
@@ -584,37 +237,6 @@ void MainWindow::update(HttpRequestWorker *worker)
     }
 }
 
-void MainWindow::on_actionSetLocalSource_triggered()
-{
-    QString textInput = QInputDialog::getText(this, tr("Set local source"), tr("source:"), QLineEdit::Normal);
-    if (textInput.isEmpty())
-        return;
-
-    QSettings settings(QSettings::IniFormat, QSettings::UserScope, "ISIS", "jv2");
-    settings.setValue("localSource", textInput);
-
-    QString msg = "If table fails to load, the local source cannot be found";
-    QMessageBox::information(this, "", msg);
-
-    QString url_str = "http://127.0.0.1:5000/setLocalSource/" + textInput.replace("/", ";");
-    HttpRequestInput input(url_str);
-    auto *worker = new HttpRequestWorker(this);
-    connect(worker, &HttpRequestWorker::on_execution_finished, [=]() { refreshTable(); });
-    worker->execute(input);
-}
-
-void MainWindow::on_actionClearLocalSource_triggered()
-{
-    QSettings settings(QSettings::IniFormat, QSettings::UserScope, "ISIS", "jv2");
-    settings.setValue("localSource", "");
-
-    QString url_str = "http://127.0.0.1:5000/clearLocalSource";
-    HttpRequestInput input(url_str);
-    auto *worker = new HttpRequestWorker(this);
-    connect(worker, &HttpRequestWorker::on_execution_finished, [=]() { refreshTable(); });
-    worker->execute(input);
-}
-
 void MainWindow::refreshTable()
 {
     for (auto i = 0; i < model_->columnCount(); ++i)
@@ -622,4 +244,29 @@ void MainWindow::refreshTable()
         ui_.runDataTable->setColumnHidden(i, true);
     }
     currentInstrumentChanged(instName_);
+}
+
+// Hide column on view menu change
+void MainWindow::columnHider(int state)
+{
+    auto *action = qobject_cast<QCheckBox *>(sender());
+
+    for (auto i = 0; i < model_->columnCount(); ++i)
+    {
+        if (action->text() == headersMap_[model_->headerData(i, Qt::Horizontal, Qt::UserRole).toString()])
+        {
+            switch (state)
+            {
+                case Qt::Unchecked:
+                    ui_.runDataTable->setColumnHidden(i, true);
+                    break;
+                case Qt::Checked:
+                    ui_.runDataTable->setColumnHidden(i, false);
+                    break;
+                default:
+                    action->setCheckState(Qt::Checked);
+            }
+            break;
+        }
+    }
 }
