+++ frontend/mainWindow.h
@@ -21,45 +21,28 @@ class MainWindow : public QMainWindow
     MainWindow(QWidget *parent = nullptr);
     ~MainWindow();
 
+    /*
+     * UI
+     */
     private:
     Ui::MainWindow ui_;
 
+    bool init_;
+    bool validSource_;
+    QPoint pos_;
+
+    /*
+     * Main Data
+     */
+    private:
     // Init
     void fillInstruments(QList<std::tuple<QString, QString, QString>> instruments);
-    void initialiseElements();
-    // Misc
-    void goToCurrentFoundIndex(QModelIndex index);                 // Selects given index
-    QList<std::tuple<QString, QString, QString>> getInstruments(); // Get Instruments from config file
-    std::vector<std::pair<QString, QString>> getFields(QString instrument, QString instType); // Get Fields from config file
     void setLoadScreen(bool state);
     QString getRunNos();
-    QDomDocument getConfig();
+
     void checkForUpdates();
 
     private slots:
-    // Search Controls
-    void updateSearch(const QString &arg1);
-    void on_actionSearch_triggered();
-    void on_actionSelectNext_triggered();
-    void on_actionSelectPrevious_triggered();
-    void on_actionSelectAll_triggered();
-    void findUp();
-    void findDown();
-    void selectAllSearches();
-    void selectIndex(QString runNumber);
-    void selectSimilar();
-    // Filter Controls
-    void on_RunFilterEdit_textChanged(const QString &arg1);
-    void on_ClearFilterButton_clicked();
-    void massSearch(QString name, QString value);
-    void on_actionMassSearchRB_No_triggered();
-    void on_actionMassSearchTitle_triggered();
-    void on_actionMassSearchUser_triggered();
-    void on_actionMassSearchRunRange_triggered();
-    void on_actionMassSearchDateRange_triggered();
-    void on_actionClear_cached_searches_triggered();
-    void goTo(HttpRequestWorker *worker, QString runNumber);
-    void on_actionRun_Number_triggered();
     // Data Selection
     void handle_result_instruments(HttpRequestWorker *worker);
     void handle_result_cycles(HttpRequestWorker *worker);
@@ -67,42 +50,13 @@ class MainWindow : public QMainWindow
     void changeCycle(QString value);
     void recentCycle();
     void changeInst(std::tuple<QString, QString, QString> instrument);
-    // Grouping
-    void on_GroupRunsButton_clicked(bool checked);
-    // Visualisation
-    void customMenuRequested(QPoint pos);
-    void handle_result_contextGraph(HttpRequestWorker *worker);
-    void contextGraph();
-    void handle_result_contextMenu(HttpRequestWorker *worker);
-    void toggleAxis(int state);
-    void getField();
-    void showStatus(qreal x, qreal y, QString title);
-
-    void handleSpectraCharting(HttpRequestWorker *worker);
-    void handleMonSpectraCharting(HttpRequestWorker *worker);
-    void plotSpectra(HttpRequestWorker *count);
-    void plotMonSpectra(HttpRequestWorker *count);
-    void getSpectrumCount();
-    void getMonitorCount();
-
-    // Normalisation options
-    void muAmps(QString runs, bool checked, QString);
-    void runDivide(QString currentDetector, QString run, bool checked);
-    void monDivide(QString currentRun, QString mon, bool checked);
 
     // Misc Interface Functions
     void removeTab(int index);
-    void savePref();
-    void clearPref();
     void columnHider(int state);
-
-    void on_actionMountPoint_triggered();
-    void on_actionClearMountPoint_triggered();
 
     void refresh(QString Status);
     void update(HttpRequestWorker *worker);
-    void on_actionSetLocalSource_triggered();
-    void on_actionClearLocalSource_triggered();
     void refreshTable();
 
     protected:
@@ -135,9 +89,95 @@ class MainWindow : public QMainWindow
     QString instDisplayName_;
     QMap<QString, QString> cyclesMap_;
     QMap<QString, QString> headersMap_;
-    // Misc
-    bool init_;
-    bool validSource_;
-    QPoint pos_;
+
+    /*
+     * Settings
+     */
+    private:
+    void savePref();
+    void clearPref();
+    // Get available instruments from config file
+    QList<std::tuple<QString, QString, QString>> getInstruments();
+    QDomDocument getConfig();
+    // Get Fields from config file
+    std::vector<std::pair<QString, QString>> getFields(QString instrument, QString instType);
+
+    private slots:
+    void on_actionMountPoint_triggered();
+    void on_actionClearMountPoint_triggered();
+    void on_actionSetLocalSource_triggered();
+    void on_actionClearLocalSource_triggered();
+
+    /*
+     * Searching
+     */
+    private:
+    void updateSearch(const QString &arg1);
+    void findUp();
+    void findDown();
+    void selectAllSearches();
+    void selectIndex(QString runNumber);
+    void selectSimilar();
+    void goToCurrentFoundIndex(QModelIndex index);
+
+    private slots:
+    void on_actionSearch_triggered();
+    void on_actionSelectNext_triggered();
+    void on_actionSelectPrevious_triggered();
+    void on_actionSelectAll_triggered();
+
+    /*
+     * Filtering
+     */
+    private:
+    void goTo(HttpRequestWorker *worker, QString runNumber);
+
+    private slots:
+    void on_RunFilterEdit_textChanged(const QString &arg1);
+    void on_ClearFilterButton_clicked();
+    void on_GroupRunsButton_clicked(bool checked);
+
+    /*
+     * Mass Search
+     */
+    private:
+    // Cached mass search results
     QList<std::tuple<HttpRequestWorker *, QString>> cachedMassSearch_;
+
+    private:
+    // Perform mass search across cycles
+    void massSearch(QString name, QString value);
+
+    private slots:
+    void on_actionMassSearchRB_No_triggered();
+    void on_actionMassSearchTitle_triggered();
+    void on_actionMassSearchUser_triggered();
+    void on_actionMassSearchRunRange_triggered();
+    void on_actionMassSearchDateRange_triggered();
+    void on_actionClear_cached_searches_triggered();
+    void on_actionRun_Number_triggered();
+
+    /*
+     * Visualisation
+     */
+    private slots:
+    void customMenuRequested(QPoint pos);
+    void handle_result_contextGraph(HttpRequestWorker *worker);
+    void contextGraph();
+    void handle_result_contextMenu(HttpRequestWorker *worker);
+    void toggleAxis(int state);
+    void getField();
+    void showStatus(qreal x, qreal y, QString title);
+
+    void handleSpectraCharting(HttpRequestWorker *worker);
+    void handleMonSpectraCharting(HttpRequestWorker *worker);
+    void plotSpectra(HttpRequestWorker *count);
+    void plotMonSpectra(HttpRequestWorker *count);
+    void getSpectrumCount();
+    void getMonitorCount();
+
+    // Normalisation options
+    void muAmps(QString runs, bool checked, QString);
+    void runDivide(QString currentDetector, QString run, bool checked);
+    void monDivide(QString currentRun, QString mon, bool checked);
 };
